social_app/requirement.txt
Django==3.0.5
django-rest-framework==0.1.0
djangorestframework==3.15.1
requests==2.27.1
psycopg2-binary==2.9.3
